Rename examples routes components to kebab-case
