+++ packages/examples/src/routes/error.tsx
