+++ packages/examples/src/routes/files-example.tsx
