+++ packages/examples/src/routes/photos-example.tsx
