+++ packages/examples/src/routes/shared-axis-example.tsx
